requirements.txt
clang==16.0.1
libclang==16.0.0
tensorflow-cpu==2.11.0
tiktoken==0.3.3
pytest==7.3.1
ruff==0.0.272
black==23.3.0
openai==0.27.8
pyyaml==6.0
